added float stuff

+++ app/src/main/java/myfirstapp/example/com/shopsuey/Stock.kt
@@ -3,8 +3,8 @@ package myfirstapp.example.com.shopsuey
 class Stock {
     var id: Int = 0
     var art_id: Int = 0
-    var stock: Int = 0
-    var minStock: Int = 0
+    var stock: Float = 0.0f
+    var minStock: Float = 0.0f
